requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
idna==2.8
multidict==4.7.3
PyYAML==5.2
yarl==1.4.2
